test(rem units): update tests for css rem units change
